pyproject.toml
[project]
name = "livekit-voice-agent"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "livekit-agents[elevenlabs,openai,silero,turn-detector,deepgram]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "mem0ai>=1.0.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "ruff>=0.14.10",
]

[tool.ruff]
# Set the maximum line length to 100 characters
line-length = 100
# Target Python version
target-version = "py313"
# Exclude common directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

[tool.ruff.lint]
# Enable a comprehensive set of rules
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
    "TCH", # flake8-type-checking
    "ARG", # flake8-unused-arguments
    "PIE", # flake8-pie
    "PT",  # flake8-pytest-style
    "RET", # flake8-return
    "RUF", # Ruff-specific rules
]
# Ignore specific rules if needed
ignore = []

[tool.ruff.lint.isort]
# Use a single line import style
force-single-line = false
# Known first party packages
known-first-party = []

[tool.ruff.format]
# Use double quotes for strings
quote-style = "double"
# Indent with spaces
indent-style = "space"
# Use spaces for indentation
skip-magic-trailing-comma = false
# Line ending style
line-ending = "auto"
